move concerns to concerns folder
