+++ app/models/concerns/stash_engine/limits.rb
@@ -1,0 +1,11 @@
+require 'active_support/concern'
+
+module StashEngine
+  module Limits
+    extend ActiveSupport::Concern
+
+    def new_upload_size_limit
+      payer_2025? || sponsored?
+    end
+  end
+end
